fixed bugs: couldnt import from strategy because of circular import: splitted strategy in two files. Strategies is where the dictionary of all strategies are + the find_strategy method, strategy_basics is where the Enum class is and the rule_based_strategy
