+++ src/strategies/macd.py
@@ -2,7 +2,7 @@ import requests
 from datetime import datetime, timedelta, timezone
 from typing import Tuple
 from config import load_api_keys
-from strategy import SideSignal
+from .strategy_basics import SideSignal
 
 def exponential_moving_average(data, period):
     if not data or period <= 0:
