+++ src/strategies/rsi.py
@@ -2,7 +2,7 @@
 import requests
 from datetime import datetime, timedelta, timezone
 from typing import Tuple
-from strategy import SideSignal
+from .strategy_basics import SideSignal
 from config import load_api_keys
 
 def calculate_rsi(closes, period: int = 14) -> float:
